Rename CouchDB storage module.
